feat(7): provide authentication for groups API
The groups API is now authenticated. There is a lot of database logic
being done in our API server right now which I feel strongly should be
offloaded to a set of PostgreSQL stored procedures, but we're not going
to have time to incorporate that before the demo.

It's not really a big deal; the application will still work just fine. I
just don't like leaving room for potential data-races.

Signed-off-by: Buckley Ross <brr322@lehigh.edu>
Change-Id: Iddd58bee7952ac34505653206e42b6de6f9503b4